# guang-source/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Guang Project - OCR transcription app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  name: string;
  /** @nullable */
  pin?: string | null;
}

export interface RecoveryInput {
  email: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
}

export interface Notebook {
  id: number;
  name: string;
  userId: number;
  noteCount: number;
  createdAt: string;
}

export interface NotebookInput {
  /** @minLength 1 */
  name: string;
}

export interface Note {
  id: number;
  title: string;
  transcribedText: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  sourceLanguage?: string | null;
  /** @nullable */
  targetLanguage?: string | null;
  /** @nullable */
  notebookId?: number | null;
  userId: number;
  createdAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  transcribedText: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  sourceLanguage?: string | null;
  /** @nullable */
  targetLanguage?: string | null;
  /** @nullable */
  notebookId?: number | null;
}

export interface NoteUpdate {
  /** @minLength 1 */
  title?: string;
  transcribedText?: string;
  /** @nullable */
  notebookId?: number | null;
}

export interface TranscribeInput {
  /** Base64-encoded image data (with or without data URI prefix) */
  imageBase64: string;
  /**
     * Target language code (e.g. "pt", "en", "es"). If null, returns original text.
     * @nullable
     */
  targetLanguage?: string | null;
  /**
     * Hint for the source language
     * @nullable
     */
  sourceLanguage?: string | null;
}

export interface TranscribeResult {
  transcribedText: string;
  detectedLanguage: string;
  /** @nullable */
  targetLanguage?: string | null;
}

export interface Language {
  code: string;
  name: string;
  nativeName: string;
}

export type GetNotesParams = {
notebookId?: number;
};