# guang-source/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Guang Project - OCR transcription app
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: notes
    description: Notes/transcriptions operations
  - name: notebooks
    description: Notebooks operations
  - name: ocr
    description: OCR transcription operations
  - name: user
    description: User profile operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/recovery:
    post:
      operationId: requestRecovery
      tags: [auth]
      summary: Request password recovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryInput"
      responses:
        "200":
          description: Recovery email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /user/me:
    get:
      operationId: getMe
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [user]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteMe
      tags: [user]
      summary: Delete current user account
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /notebooks:
    get:
      operationId: getNotebooks
      tags: [notebooks]
      summary: List user notebooks
      responses:
        "200":
          description: List of notebooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notebook"
    post:
      operationId: createNotebook
      tags: [notebooks]
      summary: Create a new notebook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotebookInput"
      responses:
        "201":
          description: Created notebook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notebook"

  /notebooks/{id}:
    get:
      operationId: getNotebook
      tags: [notebooks]
      summary: Get a notebook by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notebook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notebook"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNotebook
      tags: [notebooks]
      summary: Delete a notebook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /notes:
    get:
      operationId: getNotes
      tags: [notes]
      summary: List user notes
      parameters:
        - name: notebookId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /ocr/transcribe:
    post:
      operationId: transcribeImage
      tags: [ocr]
      summary: Transcribe text from an image using AI vision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeInput"
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscribeResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /languages:
    get:
      operationId: getLanguages
      tags: [ocr]
      summary: List supported languages
      responses:
        "200":
          description: List of supported languages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Language"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1

    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 1
        pin:
          type: ["string", "null"]

    RecoveryInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time

    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email

    Notebook:
      type: object
      required: [id, name, userId, noteCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        userId:
          type: integer
        noteCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    NotebookInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    Note:
      type: object
      required: [id, title, transcribedText, userId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        transcribedText:
          type: string
        imageUrl:
          type: ["string", "null"]
        sourceLanguage:
          type: ["string", "null"]
        targetLanguage:
          type: ["string", "null"]
        notebookId:
          type: ["integer", "null"]
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time

    NoteInput:
      type: object
      required: [title, transcribedText]
      properties:
        title:
          type: string
          minLength: 1
        transcribedText:
          type: string
        imageUrl:
          type: ["string", "null"]
        sourceLanguage:
          type: ["string", "null"]
        targetLanguage:
          type: ["string", "null"]
        notebookId:
          type: ["integer", "null"]

    NoteUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        transcribedText:
          type: string
        notebookId:
          type: ["integer", "null"]

    TranscribeInput:
      type: object
      required: [imageBase64]
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image data (with or without data URI prefix)
        targetLanguage:
          type: ["string", "null"]
          description: Target language code (e.g. "pt", "en", "es"). If null, returns original text.
        sourceLanguage:
          type: ["string", "null"]
          description: Hint for the source language

    TranscribeResult:
      type: object
      required: [transcribedText, detectedLanguage]
      properties:
        transcribedText:
          type: string
        detectedLanguage:
          type: string
        targetLanguage:
          type: ["string", "null"]

    Language:
      type: object
      required: [code, name, nativeName]
      properties:
        code:
          type: string
        name:
          type: string
        nativeName:
          type: string