# guang-source/lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Guang Project - OCR transcription app
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: number;
  title: string;
  transcribedText: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  sourceLanguage?: string | null;
  /** @nullable */
  targetLanguage?: string | null;
  /** @nullable */
  notebookId?: number | null;
  userId: number;
  createdAt: Date;
}