# guang-source/lib/api-zod/src/generated/types/transcribeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Guang Project - OCR transcription app
 * OpenAPI spec version: 0.1.0
 */

export interface TranscribeResult {
  transcribedText: string;
  detectedLanguage: string;
  /** @nullable */
  targetLanguage?: string | null;
}